feat(DateField): parse empty string as remove date
on Chrome, the date field gets an empty string as a value when the clear button is clicked. Since it is `value != null` its parsed as a date, which is an `Invalid Date`, which is a Date object. It results in an error. It should recognise the empty string value, as well as null, as a desire to set to null.